# server/schemas/openApiInterface.yml
openapi: "3.0.0"
info:
  title: Simple API overview
  version: 2.0.0
components:
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
    JWT:     # <-- arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
paths:
  /posts/allPosts:
    get:
      summary: Retrieves all posts for non authorized users
      tags:
        - Posts
      security: []
      responses:
        200:
          description: Posts retrieved
          content:
            application/json:
              schema:
                properties:
                  PostID:
                    type: integer
                  Username:
                    type: string
                  PostContent:
                    type: string
                  Created:
                    type: string
                  liked:
                    type: boolean

                  numLikes:
                    type: integer
                  numComments:
                    type: integer

                  uriGetLikes:
                    type: string
                  uriGetLikesAuth:
                    type: string
                  uriCreateLike:
                    type: string
                  canCreateLike:
                    type: boolean
                  uriDeleteLike:
                    type: string
                  canDeleteLike:
                    type: boolean

                  uriGetComments:
                    type: string
                  uriGetCommentsAuth:
                    type: string
                  uriCreateComment:
                      type: string
                  canCreateComment:
                    type: boolean

                  uriDeletePost:
                    type: string
                  canDeletePost:
                    type: boolean
              example: 
                  [
                    {
                      PostID                  : 25,
                      Username                : "user2",
                      PostContent             : "this is a new post",
                      Created                 : "2024-06-15T21:09:37.000Z",
                      liked                   : false,

                      numLikes                : 0,
                      numComments             : 6,

                      uriGetLikes             : "/likes/25",
                      uriGetLikesAuth         : "/likes/private/3",
                      uriCreateLike           : "",
                      canCreateLike           : false,
                      uriDeleteLike           : "",
                      canDeleteLike           : false,

                      uriGetComments          : "/comments/25",
                      uriGetCommentsAuth      : "/comments/private/3",
                      uriCreateComment        : "",
                      canCreateComment        : false,

                      uriDeletePost           : "/posts/25",
                      canDeletePost           : true
                    }
                  ]
        304:
          description: Posts unchanged from cache
          content:
            application/json:
              example: 
                  null
        404:
          description: No posts retrieved
  /posts/private/allPosts:
    get:
      summary: Retrieves all posts for authorized users
      tags:
        - Posts
      security: 
        - JWT: []
      responses:
        200:
          description: Posts retrieved
          content:
            application/json:
              schema:
                properties:
                  PostID:
                    type: integer
                  Username:
                    type: string
                  PostContent:
                    type: string
                  Created:
                    type: string
                  liked:
                    type: boolean

                  numLikes:
                    type: integer
                  numComments:
                    type: integer

                  uriGetLikes:
                    type: string
                  uriGetLikesAuth:
                    type: string
                  uriCreateLike:
                    type: string
                  canCreateLike:
                    type: boolean
                  uriDeleteLike:
                    type: string
                  canDeleteLike:
                    type: boolean

                  uriGetComments:
                    type: string
                  uriGetCommentsAuth:
                    type: string
                  uriCreateComment:
                      type: string
                  canCreateComment:
                    type: boolean

                  uriDeletePost:
                    type: string
                  canDeletePost:
                    type: boolean
              example: 
                  [
                    {
                      PostID             : 3,
                      Username           : "user3",
                      PostContent        : "post from user 3",
                      Created            : "2024-06-21T16:32:02.000Z",
                      ImageURL           : "https://picsum.photos/id/237/200/300",
                      liked              : false,

                      numLikes           : 0,
                      numComments        : 0,

                      uriGetLikes        : "/likes/3",
                      uriGetLikesAuth    : "/likes/private/3",
                      uriCreateLike      : "/likes/private/3/1",
                      canCreateLike      : true,
                      uriDeleteLike      : "",
                      canDeleteLike      : false,

                      uriGetComments     : "/comments/3",
                      uriGetCommentsAuth : "/comments/private/3",
                      uriCreateComment   : "/comments/private/3/1",
                      canCreateComment   : true,

                      uriDeletePost      : "/posts/private/3",
                      canDeletePost      : true
                    }
                  ]
        304:
          description: Posts unchanged from cache
          content:
            application/json:
              example: 
                  null
        401:
          description: User not authorized for that access
        404:
          description: No posts retrieved
  /posts/:post_id:
    get:
      summary: Retrieves a post for a given post_id for non authorized users
      tags:
        - Posts
      security: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Posts retrieved
          content:
            application/json:
              schema:
                properties:
                  PostID:
                    type: integer
                  Username:
                    type: string
                  PostContent:
                    type: string
                  Created:
                    type: string
                  liked:
                    type: boolean

                  numLikes:
                    type: integer
                  numComments:
                    type: integer

                  uriGetLikes:
                    type: string
                  uriGetLikesAuth:
                    type: string
                  uriCreateLike:
                    type: string
                  canCreateLike:
                    type: boolean
                  uriDeleteLike:
                    type: string
                  canDeleteLike:
                    type: boolean

                  uriGetComments:
                    type: string
                  uriGetCommentsAuth:
                    type: string
                  uriCreateComment:
                      type: string
                  canCreateComment:
                    type: boolean

                  uriDeletePost:
                    type: string
                  canDeletePost:
                    type: boolean
              example: 
                  [
                    {
                      PostID                  : 25,
                      Username                : "user2",
                      PostContent             : "this is a new post",
                      Created                 : "2024-06-15T21:09:37.000Z",
                      liked                   : false,

                      numLikes                : 0,
                      numComments             : 6,

                      uriGetLikes             : "/likes/25",
                      uriGetLikesAuth         : "/likes/private/3",
                      uriCreateLike           : "",
                      canCreateLike           : false,
                      uriDeleteLike           : "",
                      canDeleteLike           : false,

                      uriGetComments          : "/comments/25",
                      uriGetCommentsAuth      : "/comments/private/3",
                      uriCreateComment        : "",
                      canCreateComment        : false,

                      uriDeletePost           : "/posts/25",
                      canDeletePost           : true
                    }
                  ]
        304:
          description: Posts unchanged from cache
          content:
            application/json:
              example: 
                  null
        404:
          description: No posts retrieved
...